requirements.txt
flask>=2.3.0
jinja2>=3.1.0
plotly>=5.17.0
networkx>=3.0
openai>=1.12.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
numpy>=1.24.0
tiktoken>=0.5.2
requests>=2.31.0

